feat: 計算結果表示画面の実装とわりまるファイルの削除 (Issue #4)
- わけたべアプリ用の計算結果表示画面（CalculationResultScreen）を新規作成
- 料理を食べた人で割り勘計算した結果を表示する機能を実装
- 参加者別の支払い状況、精算方法、画像保存機能を含む
- App.tsxのフローを修正（料理入力 → 計算結果表示）
- わりまる関連ファイルを削除してわけたべ専用プロジェクトに変更
- Redux依存関係を削除してシンプルな状態管理に変更
- プロジェクト名を「waketabe」に変更
