+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# わりまる - リリースノート
+# わけたべ - リリースノート
 
 ## [v1.1.0] - 2024-12-01
 
@@ -75,7 +75,7 @@
 
 ### 🎉 初回リリース
 
-わりまるは、飲み会やイベントでの割り勘計算を簡単に行える Web アプリケーションです。
+わけたべは、女子会やイベントなどでいろんな料理を注文し、「誰が何を食べたか」をもとに一人あたりの支払い金額を自動計算できる Web アプリケーションです。
 
 ### ✨ 基本機能
 
@@ -121,6 +121,6 @@
 
 ---
 
-**わりまる** - 簡単・正確な割り勘計算アプリ
+**わけたべ** - 食べた分だけ、かんたん・正確に割り勘できるアプリ
 
-https://warimaru.meggumi.com
+https://waketabe.meggumi.com
