--- src/components/organisms/InputForm.tsx
@@ -1,145 +1,0 @@
-import { useDispatch, useSelector } from 'react-redux';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faUserPlus } from '@fortawesome/free-solid-svg-icons';
-import type { RootState } from '../../store/store';
-import type { AppDispatch } from '../../store/store';
-import type { Person } from '../../domain/entities';
-import {
-  addPerson,
-  deletePerson,
-  setDetailMode,
-  setNonPayingParticipants,
-} from '../../store/peopleSlice';
-import { PersonPaymentForm } from './PersonPaymentForm';
-import { Button } from '../atoms/Button';
-import { TextInput } from '../atoms/TextInput';
-
-interface InputFormProps {
-  onShowResult: () => void;
-}
-
-export const InputForm = ({ onShowResult }: InputFormProps) => {
-  const dispatch = useDispatch<AppDispatch>();
-  const people = useSelector((state: RootState) => state.people.people);
-  const isDetailMode = useSelector((state: RootState) => state.people.isDetailMode);
-  const nonPayingParticipants = useSelector((state: RootState) => state.people.nonPayingParticipants);
-
-  const handleAddPerson = () => {
-    dispatch(addPerson());
-  };
-
-  const handleDeletePerson = (personId: string) => {
-    if (window.confirm('この人物を削除してもよろしいですか？\n関連する支払い情報もすべて削除されます。')) {
-      dispatch(deletePerson(personId));
-    }
-  };
-
-  const handleShowResult = () => {
-    onShowResult();
-  };
-
-  const handleModeChange = () => {
-    if (isDetailMode) {
-      const hasMultipleRows = people.some((person: Person) => {
-        const personForm = document.querySelector(`[data-person-id="${person.id}"]`) as HTMLElement;
-        if (personForm) {
-          const inputRows = Array.from(personForm.querySelectorAll('input[data-row]')) as HTMLInputElement[];
-          return inputRows.length > 2;
-        }
-        return false;
-      });
-      
-      if (hasMultipleRows) {
-        if (!window.confirm('シンプルモードに切り替えると、入力した項目の内訳が消えます。\n合計金額のみが保持されます。\n\n続行しますか？')) {
-          return;
-        }
-      }
-    }
-    dispatch(setDetailMode(!isDetailMode));
-  };
-
-  const handleNonPayingParticipantsChange = (value: number) => {
-    dispatch(setNonPayingParticipants(value));
-  };
-
-  return (
-    <div className="space-y-4">
-      {/* 説明文を追加 */}
-      <div className="text-center text-lg font-semibold text-gray-800 bg-white/80 rounded-lg p-2 shadow-sm">
-        参加者を登録してください（ニックネームや本名どちらでもOK）
-      </div>
-      <div className="flex justify-end items-center">
-        <div className="flex items-center gap-2">
-          <span className="text-sm text-white">詳細モード</span>
-          <Button
-            onClick={handleModeChange}
-            className={`relative inline-flex h-8 w-16 items-center rounded-full transition-colors focus:outline-none focus:ring-2 focus:ring-offset-2 ${
-              isDetailMode
-                ? 'bg-blue-600 focus:ring-blue-500'
-                : 'bg-gray-300 focus:ring-gray-500'
-            }`}
-          >
-            <span
-              className={`inline-block h-6 w-6 transform rounded-full bg-white shadow-lg transition-transform ${
-                isDetailMode ? 'translate-x-9' : 'translate-x-1'
-              }`}
-            />
-          </Button>
-        </div>
-      </div>
-
-      <div className="space-y-6 bg-white/80 backdrop-blur-sm p-4 rounded-lg shadow-sm">
-        {people.map((person: Person) => (
-          <div key={person.id} className="">
-            <PersonPaymentForm
-              person={person}
-              onDeletePerson={handleDeletePerson}
-              dispatch={dispatch}
-              isDetailMode={isDetailMode}
-            />
-          </div>
-        ))}
-
-        <div className="space-y-4">
-          <Button
-            onClick={handleAddPerson}
-            className="w-full bg-blue-600 hover:bg-blue-700"
-          >
-            <FontAwesomeIcon icon={faUserPlus} className="mr-2" />
-            人物追加
-          </Button>
-        </div>
-      </div>
-
-      <div className="bg-white/80 backdrop-blur-sm p-4 rounded-lg shadow-sm">
-        <div className="flex items-center gap-2">
-          <label htmlFor="nonPayingParticipants" className="text-sm text-gray-700 font-medium">
-            支払いをしていない人数:
-          </label>
-          <TextInput
-            id="nonPayingParticipants"
-            type="number"
-            min="0"
-            max="100"
-            value={nonPayingParticipants === 0 ? '' : nonPayingParticipants}
-            onChange={e => {
-              const v = e.target.value;
-              if (v === '') return; // 空欄のときはdispatchしない
-              const num = Number(v);
-              if (!isNaN(num)) handleNonPayingParticipantsChange(num);
-            }}
-            className="w-20"
-          />
-          <span className="text-sm text-gray-700">人</span>
-        </div>
-      </div>
-
-      <Button
-        onClick={handleShowResult}
-        className="w-full bg-lime-500 hover:bg-lime-700 font-bold text-lg"
-      >
-        計算結果を見る
-      </Button>
-    </div>
-  );
-};
