--- src/domain/entities/PeopleState.ts
@@ -1,7 +1,0 @@
-import type { Person } from './Person';
-
-export interface PeopleState {
-  people: Person[];
-  isDetailMode: boolean;
-  nonPayingParticipants: number;
-}
