--- src/domain/entities/Person.ts
@@ -1,11 +1,0 @@
-export interface Person {
-  id: string;
-  name: string;
-  payments: PaymentItem[];
-}
-
-export interface PaymentItem {
-  id: string;
-  amount: number;
-  description: string;
-}
