+++ src/domain/entities/index.ts
@@ -3,6 +3,4 @@ export * from './Dish';
 export * from './Transfer';
 export * from './CalculationResult';
 export * from './ParticipantPayment';
-export * from './DishContribution';
-export * from './Person';
-export * from './PeopleState'; 
+export * from './DishContribution';
