+++ src/domain/usecases/formatCurrency.ts
@@ -1,3 +1,3 @@
 export const formatCurrency = (amount: number): string => {
-  return new Intl.NumberFormat('ja-JP').format(Math.round(amount));
+  return new Intl.NumberFormat('ja-JP').format(Math.round(amount)) + '円';
 };
